+++ views/reports.heatmap.pdf.php
@@ -1,0 +1,8 @@
+<?php
+
+/**
+ * @var CView $this
+ * @var array $data
+ */
+
+echo $data['pdf_bytes'];